Rename AddressData to AddressResult
