+++ apps/client/src/components/AddressResultStyles.js
@@ -1,7 +1,7 @@
-import styled from "styled-components";
 import { themeColor, themeSpacing } from "@datapunt/asc-ui";
+import styled from "styled-components";
 
-export const StyledAddressData = styled.div`
+export const StyledAddressResult = styled.div`
   margin-bottom: ${themeSpacing(6)};
   padding: ${themeSpacing(7)};
   border: 1px solid ${themeColor("tint", "level3")};
